requirements.txt
boto3
azure-identity
azure-mgmt-costmanagement
pyyaml
psycopg2-binary
